New User-Tweet relation

+++ app/controllers/tweets_controller.rb
@@ -1,5 +1,6 @@
 class TweetsController < ApplicationController
   def index
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
     @tweets = Tweet.all
   end
 
@@ -13,6 +14,7 @@ class TweetsController < ApplicationController
 
   def create
     @tweet = Tweet.new(tweet_params)
+    @tweet.user = current_user
 
     if @tweet.save
       redirect_to @tweet
@@ -22,7 +24,10 @@ class TweetsController < ApplicationController
   end
 
   def edit
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
     @tweet = Tweet.find(params[:id])
+
+    session_notice(:danger, 'Wrong User') unless equal_with_current_user?(@tweet.user)
   end
 
   def update
@@ -36,10 +41,16 @@ class TweetsController < ApplicationController
   end
 
   def destroy
-    tweet = Tweet.find(params[:id])
-    tweet.destroy
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
 
-    redirect_to tweets_path
+    tweet = Tweet.find(params[:id])
+
+    if equal_with_current_user?(tweet.user)
+      tweet.destroy
+      redirect_to tweets_path
+    else
+      session_notice(:danger, 'Wrong User')
+    end
   end
 
   private
